reorganize base fixtures into their own folder
