+++ tests/fixtures/state/stores/memory/__init__.py
@@ -1,0 +1,2 @@
+# ruff: noqa: F403
+from .memory import *
