+++ tests/fixtures/state/stores/rocksdb/windows.py
