+++ tests/fixtures/throwaway_1.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture()
+def my_value(request):
+    return request.param
