+++ tests/fixtures/throwaway_2.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture()
+def my_other_value(my_value):
+    return my_value * 2
