+++ tests/test_quixstreams/test_sinks/test_base/test_sink.py
@@ -1,4 +1,4 @@
-from tests.utils import DummySink
+from tests.utilities.utils import DummySink
 
 
 class TestBatchingSink:
