+++ tests/test_quixstreams/test_sinks/test_base/test_sinkmanager.py
@@ -1,5 +1,5 @@
 from quixstreams.sinks import SinkManager
-from tests.utils import DummySink
+from tests.utilities.utils import DummySink
 
 
 class TestSinkManager:
