+++ tests/test_quixstreams/test_sources/test_base/test_manager.py
@@ -6,7 +6,7 @@ import pytest
 from quixstreams.models.topics import Topic
 from quixstreams.sources import multiprocessing
 from quixstreams.sources.base.manager import SourceException, SourceManager
-from tests.utils import DummySource
+from tests.utilities.utils import DummySource
 
 
 class TestSourceManager:
