+++ tests/test_quixstreams/test_state/test_recovery/test_recovery_manager.py
@@ -9,7 +9,7 @@ from quixstreams.state.base import StorePartition
 from quixstreams.state.exceptions import InvalidStoreChangelogOffset
 from quixstreams.state.manager import SUPPORTED_STORES
 from quixstreams.state.metadata import CHANGELOG_CF_MESSAGE_HEADER
-from tests.utils import ConfluentKafkaMessageStub
+from tests.utilities.utils import ConfluentKafkaMessageStub
 
 
 class TestRecoveryManager:
